requirements.txt
asgiref==3.7.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==41.0.1
defusedxml==0.7.1
Django==4.1.10
django-allauth==0.54.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djongo==1.3.6
idna==3.4
oauthlib==3.2.2
pycparser==2.21
PyJWT==2.7.0
pymongo==3.12.3
python3-openid==3.2.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
sqlparse==0.2.4
tzdata==2023.3
urllib3==2.0.3
virtualenv==20.23.1